Refactor entire display system for Linux support
