+++ Consol3.vcxproj
@@ -20,12 +20,14 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\Consol3.cpp" />
-    <ClCompile Include="src\Display\VT24BitFrameDrawer.cpp" />
-    <ClCompile Include="src\Display\ConsoleManager.cpp" />
-    <ClCompile Include="src\Display\DitheredGreyscaleFrameDrawer.cpp" />
-    <ClCompile Include="src\Display\TextOnlyFrameDrawer.cpp" />
+    <ClCompile Include="src\Display\Linux\LinuxTerminalManager.cpp" />
+    <ClCompile Include="src\Display\Linux\LinuxTerminalManager.hpp" />
+    <ClCompile Include="src\Display\Multiplatform\VT24BitFrameDrawer.cpp" />
+    <ClCompile Include="src\Display\Windows\WindowsTerminalManager.cpp" />
+    <ClCompile Include="src\Display\Windows\DitheredGreyscaleFrameDrawer.cpp" />
+    <ClCompile Include="src\Display\Multiplatform\TextOnlyFrameDrawer.cpp" />
     <ClCompile Include="src\Display\NullFrameDrawer.hpp" />
-    <ClCompile Include="src\Display\VT8BitFrameDrawer.cpp" />
+    <ClCompile Include="src\Display\Multiplatform\VT8BitFrameDrawer.cpp" />
     <ClCompile Include="src\Engine\Rendering\AnimatedMesh.cpp" />
     <ClCompile Include="src\Engine\Rendering\AnimatedModel.cpp" />
     <ClCompile Include="src\Engine\Rendering\Shaders\PlainShader.cpp" />
@@ -38,8 +40,8 @@
     <ClCompile Include="src\Engine\Rendering\SceneRenderer.cpp" />
     <ClCompile Include="src\Engine\Rendering\Shaders\DepthMapShader.cpp" />
     <ClCompile Include="src\Engine\Rendering\Texture.cpp" />
-    <ClCompile Include="src\Display\DitheredFrameDrawer.cpp" />
-    <ClCompile Include="src\Display\GreyscaleFrameDrawer.cpp" />
+    <ClCompile Include="src\Display\Windows\DitheredFrameDrawer.cpp" />
+    <ClCompile Include="src\Display\Windows\GreyscaleFrameDrawer.cpp" />
     <ClCompile Include="src\Engine\Consol3Engine.cpp" />
     <ClCompile Include="src\Engine\Rendering\AbstractMesh.cpp" />
     <ClCompile Include="src\Engine\Rendering\Clipper.cpp" />
@@ -64,12 +66,13 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\Display\ColorMapping.hpp" />
-    <ClInclude Include="src\Display\VT24BitFrameDrawer.hpp" />
-    <ClInclude Include="src\Display\DitheredGreyscaleFrameDrawer.hpp" />
+    <ClInclude Include="src\Display\ITerminalManager.hpp" />
+    <ClInclude Include="src\Display\Multiplatform\VT24BitFrameDrawer.hpp" />
+    <ClInclude Include="src\Display\Windows\DitheredGreyscaleFrameDrawer.hpp" />
     <ClInclude Include="src\Display\IFrameDrawer.hpp" />
     <ClInclude Include="src\Display\RGBColorConstants.hpp" />
-    <ClInclude Include="src\Display\TextOnlyFrameDrawer.hpp" />
-    <ClInclude Include="src\Display\VT8BitFrameDrawer.hpp" />
+    <ClInclude Include="src\Display\Multiplatform\TextOnlyFrameDrawer.hpp" />
+    <ClInclude Include="src\Display\Multiplatform\VT8BitFrameDrawer.hpp" />
     <ClInclude Include="src\Engine\Rendering\AnimatedMesh.hpp" />
     <ClInclude Include="src\Engine\Rendering\AnimatedModel.hpp" />
     <ClInclude Include="src\Engine\Rendering\Animation.h" />
@@ -93,12 +96,12 @@
     <ClInclude Include="src\Engine\Rendering\Lighting\DirectionalLight.hpp" />
     <ClInclude Include="src\Engine\Rendering\Lighting\ILight.hpp" />
     <ClInclude Include="src\Engine\Rendering\DepthBuffer.hpp" />
-    <ClInclude Include="src\Display\DitheredFrameDrawer.hpp" />
-    <ClInclude Include="src\Display\GreyscaleFrameDrawer.hpp" />
+    <ClInclude Include="src\Display\Windows\DitheredFrameDrawer.hpp" />
+    <ClInclude Include="src\Display\Windows\GreyscaleFrameDrawer.hpp" />
     <ClInclude Include="src\Display\HSVColor.hpp" />
     <ClInclude Include="src\Display\FrameBuffer.hpp" />
     <ClInclude Include="src\Display\RGBColor.hpp" />
-    <ClInclude Include="src\Display\ConsoleManager.hpp" />
+    <ClInclude Include="src\Display\Windows\WindowsTerminalManager.hpp" />
     <ClInclude Include="src\Engine\Consol3Engine.hpp" />
     <ClInclude Include="src\Engine\Rendering\AbstractMesh.hpp" />
     <ClInclude Include="src\Engine\Rendering\Clipper.hpp" />
