+++ src/Display/RGBColorConstants.hpp
@@ -5,6 +5,9 @@
 
 namespace Display
 {
+	static const uint32_t palette_textonly[16] = { 0x000000, 0x111111, 0x212121, 0x333333, 0x444444, 0x565656, 0x666666, 0x777777,
+												   0x898989, 0x999999, 0xAAAAAA, 0xBCBCBC, 0xCCCCCC, 0xDDDDDD, 0xEFEFEF, 0xFFFFFF };
+
 	class RGBConstants
 	{
 	public:
